ref: move annotations to governance store and epoch store
