+++ stores/epoch/src/main/java/com/bloxbean/cardano/yaci/store/epoch/annotation/LocalEpochParam.java
@@ -1,4 +1,4 @@
-package com.bloxbean.cardano.yaci.store.core.annotation;
+package com.bloxbean.cardano.yaci.store.epoch.annotation;
 
 import org.springframework.context.annotation.Conditional;
 
@@ -7,6 +7,6 @@ import java.lang.annotation.*;
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Conditional(LocalGovStateCondition.class)
-public @interface LocalGovState {
+@Conditional(LocalEpochParamCondition.class)
+public @interface LocalEpochParam {
 }
