+++ stores/governance/src/main/java/com/bloxbean/cardano/yaci/store/governance/annotation/LocalGovState.java
@@ -1,4 +1,4 @@
-package com.bloxbean.cardano.yaci.store.core.annotation;
+package com.bloxbean.cardano.yaci.store.governance.annotation;
 
 import org.springframework.context.annotation.Conditional;
 
@@ -7,6 +7,6 @@ import java.lang.annotation.*;
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Conditional(LocalEpochParamCondition.class)
-public @interface LocalEpochParam {
+@Conditional(LocalGovStateCondition.class)
+public @interface LocalGovState {
 }
